requirements.txt
pandas == 2.0.3
mlflow == 2.7.0
scikit-learn == 1.3.0
zipp == 3.16.2
requests == 2.31.0
xgboost == 2.0.0
# keras == 2.13.1
# tensorflow == 2.13.1
numpy == 1.24.3
# apache-airflow-providers-papermill == 3.4.0
# papermill[all] == 2.4.0
# scrapbook[all] == 0.5.0
openpyxl == 3.1.2
apache-airflow[amazon]
boto3